requirements.txt
av==12.2.0
joblib==1.4.2
matplotlib==3.9.1
mediapipe==0.10.14
numpy==1.26.0
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
streamlit==1.36.0
streamlit_webrtc==0.47.7
